# Task11.1HD/Prototypes and Testing/smart_clothesline_backend_mqtt/smart_clothesline_backend_mqtt/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Smart Clothesline Robot MQTT Backend API",
    "version": "2.0.0",
    "description": "Flask dashboard/backend for Smart Clothesline Robot using MQTT broker as Nano/Pi middleman."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current backend"
    }
  ],
...
  "components": {
    "schemas": {
      "CommandRequest": {
        "type": "object",
        "required": [
          "command"
        ],
        "properties": {
          "command": {
            "type": "string",
            "enum": [
              "MOVE_TO_SAFE",
              "RETURN_TO_DRYING",
              "STOP",
              "SET_ACTIVE_TRUE",
              "SET_ACTIVE_FALSE",
              "NONE"
            ],
            "example": "MOVE_TO_SAFE"
          }
        }
      },
      "TelemetryRequest": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "state": {
            "type": "string",
            "example": "DRYING"
          },
          "target": {
            "type": "string",
            "example": "NONE"
          },
          "event": {
            "type": "string",
            "example": "SUN_BALANCED"
          },
          "line_pattern": {
            "type": "string",
            "example": "010"
          },
          "rain_raw": {
            "type": "integer",
            "example": 4020
          },
          "rain_status": {
            "type": "string",
            "example": "DRY"
          },
          "left_lux": {
            "type": "number",
            "example": 610.5
          },
          "right_lux": {
            "type": "number",
            "example": 702.1
          },
          "laundry_active": {
            "type": "boolean",
            "example": true
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/telemetry": {
      "post": {
        "tags": [
          "Telemetry"
        ],
        "summary": "Legacy HTTP telemetry test endpoint",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Telemetry accepted"
          },
          "400": {
            "description": "Invalid JSON"
          }
        }
      }
    },
...
    "/api/command": {
      "post": {
        "tags": [
          "Commands"
        ],
        "summary": "Publish command to Nano over MQTT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command published"
          },
          "400": {
            "description": "Invalid command or MQTT unavailable"
          }
        }
      }
    },
...
    "/api/logs": {
      "get": {
        "tags": [
          "Logs"
        ],
        "summary": "Recent telemetry logs",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recent logs"
          }
        }
      }
    },
...
  }
}